guiao3 SD kiko
INCOMPLETE BUT WORKING

+++ 3º Ano/1º Semestre/SD/2018_19/guioes/guiao3/kiko (incomplete but working)/Conta.java
@@ -1,10 +1,10 @@
 import java.util.concurrent.locks.ReentrantLock;
 
 public class Conta {
-    private int saldo;
+    private double saldo;
     private ReentrantLock lockConta = new ReentrantLock();
 
-    public Conta(int saldo){
+    public Conta(double saldo){
         this.saldo = saldo;
     }
